Replace PersonProvider with hook instead
