+++ src/scenes/AcceleratorRouter/People/index.tsx
@@ -6,29 +6,19 @@ import { APP_PATHS } from 'src/constants';
 import EditView from './EditView';
 import ListView from './ListView';
 import NewView from './NewView';
-import PersonProvider from './PersonProvider';
 import SingleView from './SingleView';
 
 const PeopleRouter = () => {
   return (
     <Routes>
-      <Route
-        path={'new'}
-        element={
-          <PersonProvider>
-            <NewView />
-          </PersonProvider>
-        }
-      />
+      <Route path={'new'} element={<NewView />} />
       <Route
         path={':userId/*'}
         element={
-          <PersonProvider>
-            <Routes>
-              <Route path={''} element={<SingleView />} />
-              <Route path={'edit'} element={<EditView />} />
-            </Routes>
-          </PersonProvider>
+          <Routes>
+            <Route path={''} element={<SingleView />} />
+            <Route path={'edit'} element={<EditView />} />
+          </Routes>
         }
       />
       <Route path={''} element={<ListView />} />
